requirements.txt
ldap3
requests
lxml
python-docx==0.8.7